# docs/reference/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "DeleteMeetingResp": {
        "description": "Resposta de `DELETE /api/v1/meetings/{id}`.",
        "properties": {
          "ok": {
            "description": "Sempre `true`.",
            "type": "boolean"
          },
          "room_deleted": {
            "description": "`false` quando a sala foi mantida por ter gravações (ou a reunião não\ntinha sala).",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "room_deleted"
        ],
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "Envelope de erro de todas as superfícies (ver `error.rs`).",
        "properties": {
          "code": {
            "description": "Código estável (contrato), p.ex. `registration.domain_taken`.",
            "type": "string"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/FieldViolation"
            },
            "type": "array"
          },
          "error": {
            "description": "Mensagem para pessoas. Pode mudar.",
            "type": "string"
          },
          "request_id": {
            "description": "Igual ao cabeçalho `X-Request-Id`.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "error",
          "code",
          "details"
        ],
        "type": "object"
      },
...
      "FieldViolation": {
        "properties": {
          "description": {
            "type": "string"
          },
          "field": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "description"
        ],
        "type": "object"
      },
...
      "InviteeReq": {
        "properties": {
          "email": {
            "type": "string"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "email"
        ],
        "type": "object"
      },
      "InviteeResp": {
        "properties": {
          "email": {
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "email",
          "user_id"
        ],
        "type": "object"
      },
...
      "MeetingResp": {
        "properties": {
          "duration_min": {
            "format": "int32",
            "type": "integer"
          },
          "existing": {
            "description": "`true` quando o pedido reencontrou uma reunião já existente pela\n`external_ref` em vez de criar uma nova.",
            "type": "boolean"
          },
          "external_ref": {
            "type": [
              "string",
              "null"
            ]
          },
          "host_email": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "invitees": {
            "description": "Convidados efectivamente registados (email -> id do utilizador).",
            "items": {
              "$ref": "#/components/schemas/InviteeResp"
            },
            "type": "array"
          },
          "join_url": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "room_code": {
            "type": "string"
          },
          "skipped": {
            "description": "Emails NÃO registados e porquê. Nunca falha o pedido por causa de um\nconvidado — o link tem de sair na mesma.",
            "items": {
              "$ref": "#/components/schemas/SkippedInvitee"
            },
            "type": "array"
          },
          "starts_at": {
            "format": "date-time",
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "starts_at",
          "duration_min",
          "kind",
          "room_code",
          "join_url",
          "host_email",
          "invitees",
          "skipped",
          "existing"
        ],
        "type": "object"
      },
...
      "PatchMeetingReq": {
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "duration_min": {
            "format": "int32",
            "type": [
              "integer",
              "null"
            ]
          },
          "invitees": {
            "description": "Se presente, **substitui** o conjunto de convidados. O estado de\nresposta (aceite/recusado) de quem se mantém é preservado.",
            "items": {
              "$ref": "#/components/schemas/InviteeReq"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "starts_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
      "SkippedInvitee": {
        "properties": {
          "email": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "reason"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "description": "Chave de API da organização (`dlx_…`).",
        "scheme": "bearer",
        "type": "http"
      },
      "session": {
        "bearerFormat": "JWT",
        "description": "Access token da sessão (15 min), renovado por /api/auth/refresh.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Superfície estável do inquilino para SDK, mobile e integrações. Chave `dlx_` com escopos. Muda só com v2.",
    "license": {
      "name": ""
    },
    "title": "Delonix Meet — API pública v1",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/meetings/{meeting_id}": {
      "delete": {
        "description": "A sala só é apagada se não tiver gravações: a reunião pode ter acontecido\ne sido cancelada no calendário depois, e uma gravação é um artefacto que o\nutilizador espera manter (a FK `recordings.room_id` é ON DELETE CASCADE —\napagar a sala apagaria o registo da gravação).",
        "operationId": "delete",
        "parameters": [
          {
            "description": "Id da reunião",
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteMeetingResp"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "chave de API ausente, inválida ou revogada (`auth.unauthenticated`), ou expirada (`api_key.expired`)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "a chave não tem o escopo `meetings:write` (`api_key.scope_missing`, escopo em `details`)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "a reunião não existe ou o dono não é membro da organização da chave"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "rate-limit da v1, por chave (`Retry-After` com o que falta da janela)"
          }
        },
        "security": [
          {
            "api_key": [
              "meetings:write"
            ]
          }
        ],
        "summary": "`DELETE /api/v1/meetings/{id}` — cancela a reunião e liberta a sala.",
        "tags": [
          "meetings"
        ]
      },
      "get": {
        "operationId": "get_one",
        "parameters": [
          {
            "description": "Id da reunião",
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingResp"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "a chave não tem o escopo `meetings:read`"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "não existe ou é de outra organização"
          }
        },
        "security": [
          {
            "api_key": [
              "meetings:read"
            ]
          }
        ],
        "summary": "`GET /api/v1/meetings/{meeting_id}` — uma reunião da organização da chave,\nna mesma forma que a criação devolve.",
        "tags": [
          "meetings"
        ]
      },
      "patch": {
        "operationId": "patch",
        "parameters": [
          {
            "description": "Id da reunião",
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchMeetingReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingResp"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "`title` vazio, `duration_min` fora de 1-1440 ou mais de 200 convidados"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "chave de API ausente, inválida ou revogada (`auth.unauthenticated`), ou expirada (`api_key.expired`)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "a chave não tem o escopo `meetings:write` (`api_key.scope_missing`, escopo em `details`)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "a reunião não existe ou o dono não é membro da organização da chave"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "rate-limit da v1, por chave (`Retry-After` com o que falta da janela)"
          }
        },
        "security": [
          {
            "api_key": [
              "meetings:write"
            ]
          }
        ],
        "summary": "`PATCH /api/v1/meetings/{id}` — reagendar / renomear / actualizar convidados.",
        "tags": [
          "meetings"
        ]
      }
    },
...
  }
}